renamed enum
