+++ rule-engine/rule-engine-components/src/main/java/org/thingsboard/rule/engine/transform/OriginatorSource.java
@@ -15,7 +15,7 @@
  */
 package org.thingsboard.rule.engine.transform;
 
-public enum ChangeOriginatorSource {
+public enum OriginatorSource {
     CUSTOMER,
     TENANT,
     RELATED,
